chore: Update README, CHANGELOG, examples.
Various user faces docs have been updated, and the default target format
has been changed to int based on my personal testing seeming to indicate
that int is more common.

+++ src/config/audio.rs
@@ -81,7 +81,7 @@ impl Audio {
     pub fn sample_format(&self) -> Result<SampleFormat, Box<dyn Error>> {
         match self.sample_format.as_deref() {
             Some(format) => SampleFormat::from_str(format),
-            None => Ok(SampleFormat::Float), // Default to float for better precision
+            None => Ok(SampleFormat::Int),
         }
     }
